release 3.1.180 source code

+++ aad/include/huaweicloud/aad/v2/model/Curve.h
@@ -59,10 +59,10 @@ public:
     /// 时间戳
     /// </summary>
 
-    int32_t getTime() const;
+    int64_t getTime() const;
     bool timeIsSet() const;
     void unsettime();
-    void setTime(int32_t value);
+    void setTime(int64_t value);
 
 
 protected:
@@ -70,7 +70,7 @@ protected:
     bool inIsSet_;
     float out_;
     bool outIsSet_;
-    int32_t time_;
+    int64_t time_;
     bool timeIsSet_;
 
 };
